requirements.txt
# Image processing library (equivalent to Sharp in Node.js)
Pillow>=10.0.0

# AWS S3 library for uploading photos
boto3>=1.26.0

# Google Drive API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Data processing
pandas>=1.5.0

# CLI framework
typer>=0.9.0

# Configuration management
PyYAML>=6.0

# Rich CLI formatting and progress bars
rich>=13.0.0